requirements.txt
PyQt5==5.15.9
Pillow>=9.0.0
qt-material==2.17